requirements.txt
azure-identity
azure-keyvault
azure-keyvault-secrets
azure-storage-blob
psycopg2-binary
pyodbc
redis